[FIX] Fixing condition handling
Fixed condition handling so that it should work for both ``BehavioralPLS`` and
``MeanCeneteredPLS``. There's an unfortunate amount of variable passing that
happens now between the ``BasePLS`` class and the classes in types.py, but that
can be fixed in later commits.

Need to add tests comparing results of PLS analyses in Python to "ground truth"
Matlab-based results.