[FIX] Fixes permutations in behavioral PLS
For some reason (?), in the Matlab implementation of behavioral PLS, the
`Y` array (behavior) is permuted during permutation testing instead of
the `X` array. This yields VASTLY different permutation results when
comparing between Matlab and Python (even though it seems like it
shouldn't matter?), so this commit fixes this error.