[TEST] Fixes Matlab compatibility test issues
Changes to the bootstrapping memory errors resulted in failure of tests
ensuring backwards compatibility with Matlab (thanks to differences in
split-half resampling results). These changes update how we are
comparing functional equivalence of results (not numerically equivalent)
when we cannot correlate the results (i.e., len(x) <= 2). We are now
assessing whether the sign is consistent between the result vectors and
that the absolute difference between the results are within a tolerance
(`ftol`).